global: add coderabbit configuration file (#838)
This PR adds a configuration file for coderabbit to the repo. Most of the settings are pulled from the coderabbit configuration command, with a few changes:

- `collapse_walkthrough: true` (previously `false`) [docs](https://docs.coderabbit.ai/reference/configuration#param-collapse-walkthrough)
- `sequence_diagrams: false` (previously `true`) [docs](https://docs.coderabbit.ai/reference/configuration#param-sequence-diagrams)
- `path_filters: ['!vendor/']` (previously `[]`) [docs](https://docs.coderabbit.ai/reference/configuration#param-path-filters)
- `github-checks.timeout_ms: 900000` (previously `90000`) [docs](https://docs.coderabbit.ai/reference/configuration#param-timeout-ms)

The first two are to reduce the noise from the coderabbit comment with the walkthrough and diagrams. The diagrams especially do not seem very useful and take up a lot of space.

The third option is to exclude the vendor directory; we cannot change it so there is no reason to review it.

The fourth option bumps the timeout on github checks to 15 minutes from 90 seconds. Since the linter workflow takes 6-7 minutes to complete, coderabbit would previously not have access to these checks.